Fix udf implicit cast & update doc
